+++ chowpronto_frontend/src/api/fetchPreviousOrders.js
@@ -1,0 +1,17 @@
+function fetchPreviousOrders(token, _id) {
+  return fetch(`/api/orders/patron/${_id}`, {
+    method: "GET",
+    headers: {
+      Authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+  }).then((res) => {
+    if (res.status >= 200 && res.status < 300) {
+      return res.json();
+    } else {
+      throw res;
+    }
+  });
+}
+
+export default fetchPreviousOrders;
