+++ chowpronto_frontend/src/hooks/useAuth.js
@@ -9,7 +9,6 @@ function useAuth() {
     const storageData = await getFromStorage();
     if (storageData && storageData.length > 0) {
       const userDetails = await getUserById(storageData);
-      console.log(userDetails);
       if (userDetails && userDetails.patron) {
         setUserDetailsToContext({ ...userDetails, token: storageData });
       } else {
