+++ chowpronto_frontend/src/hooks/usePreviousOrders.js
@@ -1,35 +1,24 @@
 import { useEffect, useState, useContext } from "react";
 import UserContext from "../state/UserContext";
+import fetchPreviousOrders from "../api/fetchPreviousOrders";
 
-function usePreviousOrders() {
+function usePreviousOrders(setErrMsg) {
   const { user } = useContext(UserContext);
   const [orders, setOrders] = useState(null);
-  useEffect(() => {
-    fetchPreviousOrders(user.token, user.patron._id)
-      .then((data) => setOrders(data))
-      .catch((err) =>
-        err.json().then((json) => {
-          console.log(json.errorMsg);
-        })
-      );
-  }, [user.token, user.patron._id]);
-  return orders;
-}
+  const patronId =
+    user && user.patron && user.patron._id ? user.patron._id : null;
+  const token = user && user.token ? user.token : null;
 
-function fetchPreviousOrders(token, _id) {
-  return fetch(`/api/orders/patron/${_id}`, {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${token}`,
-      "Content-Type": "application/json",
-    },
-  }).then((res) => {
-    if (res.status >= 200 && res.status < 300) {
-      return res.json();
-    } else {
-      throw res;
+  useEffect(() => {
+    if (token && patronId) {
+      fetchPreviousOrders(token, patronId)
+        .then((data) => setOrders(data))
+        .catch((err) =>
+          setErrMsg("Can't retrieve info about previous orders from server")
+        );
     }
-  });
+  }, [patronId, token, setErrMsg]);
+  return orders;
 }
 
 export default usePreviousOrders;
